isolation of the new IT tests of the post api
